add GlobalExceptionHandler and Scheduler

+++ src/main/java/com/example/shop/model/ProductDto.java
@@ -1,11 +1,13 @@
 package com.example.shop.model;
+
 import lombok.Builder;
+import lombok.Data;
 import lombok.Value;
 import org.springframework.lang.Nullable;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
-
 @Value
 @Builder
 public class ProductDto {
@@ -19,7 +21,7 @@ public class ProductDto {
     private final String description;
     @Nullable
     private final Category category;
-    private final Double price;
+    private final BigDecimal price;
     private Integer quantity;
     private final LocalDateTime lastQuantityChange;
     private final LocalDateTime createDate;
